aggregator: split out walker
